DRY up loading fonts for Native and Web
