+++ components/Main.js
@@ -21,7 +21,6 @@ import Listing from '../screens/Listing';
 import Menu from '../components/Menu';
 import ImageGallery from '../components/ImageGallery';
 import {useStateValue} from "../components/State";
-import { useFonts } from 'expo-font';
 import { getStyles, getImage } from '../utils';
 import { Dimensions, Platform } from 'react-native';
 
@@ -51,30 +50,6 @@ function Main(props) {
       dispatch({type: 'setDimensions', value: set_to})
     }
   };
-
-  let fonts = {
-      'ApercuMedium': {
-          uri: require('../public/fonts/ApercuRegular.ttf'),
-          display: 'swap',
-      },
-      'ApercuLight': {
-          uri: require('../public/fonts/ApercuLight.ttf'),
-          display: 'swap',
-      },
-      'KnockoutBold': {
-          uri: require('../public/fonts/Knockout_HTF71-FullMiddlewt_Regular.otf'),
-          display: 'swap',
-      },
-      'KnockoutWelterWeight': {
-          uri: require('../public/fonts/Knockout_HTF50-Welterweight_Regular.otf'),
-          display: 'swap',
-      },
-      'KnockoutFeatherWeight': {
-          uri: require('../public/fonts/Knockout_HTF48-Featherweight_Regular.otf'),
-          display: 'swap',
-      }
-  };
-  const [ fontsReady, fontsError ] = useFonts(fonts);
 
   useEffect(() => {
     Dimensions.addEventListener("change", onChangeScreenSize);
@@ -117,11 +92,6 @@ function Main(props) {
         window.removeEventListener('scroll', scrollEventListener, false)
       }
     }, []);
-  }
-
-  if (!isWeb && !fontsReady) {
-    console.log('render loading first');
-    return ( <Text>Loading...</Text> )
   }
 
   function renderContent(props) {
